Moved UI classes to their respective packages to separate them by screen
