+++ app/src/main/java/com/example/flickrgallery/ui/explore/ExploreBindingAdapter.kt
@@ -1,10 +1,10 @@
-package com.example.flickrgallery.ui
+package com.example.flickrgallery.ui.explore
 
 import android.view.View
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.flickrgallery.model.Photo
-
+import com.example.flickrgallery.ui.common.PhotosAdapter
 
 
 @BindingAdapter("items")
