+++ app/src/main/java/com/example/flickrgallery/ui/storedLocations/StoredLocationsViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.flickrgallery.ui
+package com.example.flickrgallery.ui.storedLocations
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
