+++ app/src/main/res/layout/saved_photos_fragment.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.example.flickrgallery.ui.SavedPhotosViewModel" />
+            type="com.example.flickrgallery.ui.savedPhotos.SavedPhotosViewModel" />
     </data>
 
     <LinearLayout
