+++ app/src/main/res/layout/stored_location_fragment.xml
@@ -6,14 +6,14 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.flickrgallery.ui.StoredLocationViewModel" />
+            type="com.example.flickrgallery.ui.storedLocationPhotos.StoredLocationViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:theme="@style/Theme.MaterialComponents.DayNight"
-        tools:context=".ui.ExploreFragment">
+        tools:context=".ui.explore.ExploreFragment">
 
 
         <androidx.recyclerview.widget.RecyclerView
