+++ app/src/main/res/layout/stored_locations_fragment.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.flickrgallery.ui.StoredLocationsViewModel" />
+            type="com.example.flickrgallery.ui.storedLocations.StoredLocationsViewModel" />
     </data>
 
     <LinearLayout
